+++ packages/ui/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The following is a curated list of changes in the Enact ui module, newest changes on the top.
 
+## [unreleased]
+
+### Deprecated
+
+- `ui/Holdable` and `ui/Pressable` to be replaced by `ui/Touchable` in 2.0.0
+
 ## [1.13.3] - 2017-01-16
 
 No significant changes.